pyproject.toml
[project]
name = "node0"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = "Apache-2.0"
license-files = ["*LICEN[CS]E*"]
dependencies = [
    "PyYAML",
    "prometheus_client",
    "torch==2.7.0",
    "numpy>=1.17",
    "scipy>=1.2.1",
    "prefetch_generator>=1.0.1",
    "msgpack>=0.5.6",
    "sortedcontainers",
    "uvloop>=0.14.0",
    "grpcio-tools>=1.33.2",
    "protobuf>=5.29.0",
    "configargparse>=1.2.3",
    "py-multihash>=0.2.3",
    "cryptography>=3.4.6",
    "pydantic>=2.0.0",
    "packaging>=20.9",
    "varint>=1.0.2",
    "base58>=1.0.2",
    "netaddr>=1.3.0",
    "idna>=3.10",
    "py-cid>=0.3.0",
    "requests>=2.32.3",
    "speedtest-cli>=2.1.3",
    "psutil>=7.0.0",
    "hivemind @ git+https://github.com/learning-at-home/hivemind.git@4d5c41495be082490ea44cce4e9dd58f9926bb4e"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
